Moved ULUtil into users package
